refactor(logs): extract shared LogRow component
Pull the inlined row layout (timestamp / level pill / endpoint /
message-or-ToolCallRow) out of `RelayLogs.svelte` into a new pure
presentation component `LogRow.svelte`. Same DOM, same Tailwind
classes, same `endpointStripeStyle` — the refactor is intentionally
visual-no-op.

Prep for Slice D.2: the live-streaming LogsTab will render the same
`ParsedLogLine` shape via the same component, satisfying the spec's
"identical rendering across views" requirement.
